+++ docs/docs/vulnerability/detection/data-source.md
@@ -37,7 +37,7 @@
 | Rust     | [Open Source Vulnerabilities (crates.io)][rust-osv] |       ✅        |     -     |
 | .NET     | [GitHub Advisory Database (NuGet)][dotnet-ghsa]     |       ✅        |     -     |
 | C/C++    | [GitLab Advisories Community][gitlab]               |       ✅        |  1 month  |
-| Dark     | [GitHub Advisory Database (Pub)][pub-ghsa]          |       ✅        |     -     |
+| Dart     | [GitHub Advisory Database (Pub)][pub-ghsa]          |       ✅        |     -     |
 
 [^1]: Intentional delay between vulnerability disclosure and registration in the DB
