+++ docs/docs/vulnerability/detection/language.md
@@ -27,7 +27,7 @@
 | Rust     | Cargo.lock                                                                                 |     ✅     |     ✅      |        ✅        |        ✅        | included         |            -             |
 |          | Binaries built with [cargo-auditable](https://github.com/rust-secure-code/cargo-auditable) |     ✅     |     ✅      |        -        |        -        | excluded         |            -             |
 | C/C++    | conan.lock[^13]                                                                            |     -     |     -      |        ✅        |        ✅        | excluded         |            -             |   
-| Dark     | pubspec.lock                                                                               |     ✅     |     ✅      |        -        |        -        | excluded         |            -             |
+| Dart     | pubspec.lock                                                                               |     ✅     |     ✅      |        -        |        -        | excluded         |            -             |
 
 The path of these files does not matter.
