+++ pkg/fanal/analyzer/language/dart/pubspeclock/testdata/happy.lock
